Safari-611.1.10/LayoutTests/accessibility/gtk/aria-haspopup-expected.txt

This tests the exposure of aria-haspopup

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



test1 hasPopup: false
AXPlatformAttributes: computed-role:button, xml-roles:button, html-id:test1, tag:div, toolkit:WebKitGtk

test2 hasPopup: true
AXPlatformAttributes: computed-role:button, xml-roles:button, haspopup:true, html-id:test2, tag:div, toolkit:WebKitGtk

test3 hasPopup: false
AXPlatformAttributes: computed-role:button, xml-roles:button, haspopup:false, html-id:test3, tag:div, toolkit:WebKitGtk

test4 hasPopup: true
AXPlatformAttributes: computed-role:button, xml-roles:button, haspopup:dialog, html-id:test4, tag:div, toolkit:WebKitGtk

test5 hasPopup: true
AXPlatformAttributes: computed-role:button, xml-roles:button, haspopup:grid, html-id:test5, tag:div, toolkit:WebKitGtk

test6 hasPopup: true
AXPlatformAttributes: computed-role:button, xml-roles:button, haspopup:listbox, html-id:test6, tag:div, toolkit:WebKitGtk

test7 hasPopup: true
AXPlatformAttributes: computed-role:button, xml-roles:button, haspopup:menu, html-id:test7, tag:div, toolkit:WebKitGtk

test8 hasPopup: true
AXPlatformAttributes: computed-role:button, xml-roles:button, haspopup:tree, html-id:test8, tag:div, toolkit:WebKitGtk

test9 hasPopup: false
AXPlatformAttributes: computed-role:button, xml-roles:button, haspopup:false, html-id:test9, tag:div, toolkit:WebKitGtk

test10 hasPopup: false
AXPlatformAttributes: computed-role:button, xml-roles:button, haspopup:false, html-id:test10, tag:div, toolkit:WebKitGtk

test11 hasPopup: true
AXPlatformAttributes: computed-role:combobox, xml-roles:combobox, readonly:false, haspopup:listbox, html-id:test11, tag:div, toolkit:WebKitGtk

test12 hasPopup: true
AXPlatformAttributes: computed-role:combobox, xml-roles:combobox, readonly:false, haspopup:true, html-id:test12, tag:div, toolkit:WebKitGtk

test13 hasPopup: false
AXPlatformAttributes: computed-role:combobox, xml-roles:combobox, readonly:false, haspopup:false, html-id:test13, tag:div, toolkit:WebKitGtk

test14 hasPopup: true
AXPlatformAttributes: computed-role:combobox, xml-roles:combobox, readonly:false, haspopup:dialog, html-id:test14, tag:div, toolkit:WebKitGtk

test15 hasPopup: true
AXPlatformAttributes: computed-role:combobox, xml-roles:combobox, readonly:false, haspopup:grid, html-id:test15, tag:div, toolkit:WebKitGtk

test16 hasPopup: true
AXPlatformAttributes: computed-role:combobox, xml-roles:combobox, readonly:false, haspopup:listbox, html-id:test16, tag:div, toolkit:WebKitGtk

test17 hasPopup: true
AXPlatformAttributes: computed-role:combobox, xml-roles:combobox, readonly:false, haspopup:menu, html-id:test17, tag:div, toolkit:WebKitGtk

test18 hasPopup: true
AXPlatformAttributes: computed-role:combobox, xml-roles:combobox, readonly:false, haspopup:tree, html-id:test18, tag:div, toolkit:WebKitGtk

test19 hasPopup: false
AXPlatformAttributes: computed-role:combobox, xml-roles:combobox, readonly:false, haspopup:false, html-id:test19, tag:div, toolkit:WebKitGtk

test20 hasPopup: true
AXPlatformAttributes: computed-role:combobox, xml-roles:combobox, readonly:false, haspopup:listbox, html-id:test20, tag:div, toolkit:WebKitGtk
PASS successfullyParsed is true

TEST COMPLETE
